Add services object to contain moving parts of the app
