+++ frontend/src/app/ui/app/App.test.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 import { render } from "@testing-library/react";
 import { App } from "./App";
+import { createServices } from "app/services/services";
 
 test("renders learn react link", () => {
-  const { getByText } = render(<App />);
-  const linkElement = getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+  const services = createServices();
+  const res = render(<App services={services} />);
+  const el = res.getByText(/Tracking/i);
+  expect(el).toBeInTheDocument();
 });
