+++ frontend/src/app/utils/routeUtils.tsx
@@ -1,24 +1,10 @@
 import { generatePath } from "react-router";
-import qhistory from "qhistory";
-import { parse, stringify } from "query-string";
-import { createBrowserHistory } from "history";
+import { stringify } from "query-string";
 
 export const routes = {
   BOOKING: "/booking",
   TRACKING: "/tracking",
 };
-
-const stringifyWithBrackets = (params: object) =>
-  stringify(params, { arrayFormat: "bracket" });
-
-const parseWithBrackets = (url: string) =>
-  parse(url, { arrayFormat: "bracket" });
-
-export const history = qhistory(
-  createBrowserHistory(),
-  stringifyWithBrackets,
-  parseWithBrackets
-);
 
 interface IUrlParams {
   [paramName: string]: string | number | boolean;
